Added methods for processing person subscriptions
Process, subscribe and unsubscribe methods for persons have been
added. Process method is implemented but needs adaptation to work
properly (addModule or updateModule method params are not compatible
yet). Process will read id_list and manage which persons are already
subscribed and which not. It will also transform coordinates and
heading to omnetpp compatible stuff.