PersonSubscriptionManager added
This is intended to encapsulate everything related to person subscription.
Currently it is only able to run in a polling mode since subscriptions for
persons are not working yet.

Call update with a list of ids to update the internal state. And then use
getters to receive information about new and disappeared persons.